Separate flux into constants and others
